Split Interfaces.ts contents All interfacesare now located in separate files.
