+++ src/iclaim_service.ts
@@ -1,0 +1,5 @@
+import {Claim, HasClaim} from './type_aliases';
+
+export interface IClaimService {
+  hasClaim(roles: Array<string>, claim: Claim, namespace?: string): HasClaim;
+}
