+++ src/index.ts
@@ -1,1 +1,7 @@
-export * from './interfaces';
+export * from './iauth_service';
+export * from './iclaim_service';
+export * from './iidentity_service';
+export * from './isession_store_entity';
+export * from './itoken_adapter';
+export * from './itoken_service';
+export * from './type_aliases';
