+++ src/isession_store_entity.ts
@@ -1,0 +1,7 @@
+import {IEntity} from '@essential-projects/core_contracts';
+
+export interface ISessionStoreEntity extends IEntity {
+  identityId: string;
+  systemUserId: string;
+  data: any;
+}
