+++ src/type_aliases.ts
@@ -1,0 +1,2 @@
+export type Claim = string;
+export type HasClaim = boolean;
